+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
-- Support for C#'s `*.packages.config` lockfile type
+- Support for C#'s `packages.*.config` lockfile type
 
 ## 7.1.5 - 2024-11-26